+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [6.7.7] - 2025-14-05
+
+### Added
+
+- lookup_compound() now allows input in upper/lower case.
+- lookup_compound() with new `preserve_case` parameter : Whether to preserve the letter case from input to suggestion.
+- transfer_case doesn't transfer lower case of whitespace in source to non-whitespace char in target.
+
 ## [6.7.6] - 2025-14-05
 
 ### Added